Avoid false-positive TXB0 header check
This could happens if the first frame to be written is a standard data
frame with identifier 0 and no data. In this case the TXB0 header would
be all zero. In this case the header would not be written und result in
sending an undefined frame, because the TXB0 registers are undefined
after a reset.